MOD v0.0.10

+++ src/application/classes/ModulePreset.coffee
@@ -24,36 +24,6 @@ ZookeeperHandler = require("./ZookeeperHandler")
 #
 class ModulePreset
 
-  # filesystemhandler
-  @fs: require("q-io/fs")
-
-  # read preset from filesystem
-  #
-  # @return [Promise] resolves with array of filecontents
-  #
-  @readPresetFromFS: =>
-    self = @
-    deferred = Q.defer()
-    self.fs.listTree(process.env.MCONN_PRESET_DIR)
-    .then (result) ->
-      result.splice(0,1) #remove root folder in array
-      fileContents = []
-      async.eachLimit(
-        result, 1
-      ,(path, done) ->
-        self.fs.read(path)
-        .then (content) ->
-          fileContents.push(JSON.parse(content))
-        .catch (error) ->
-          logger.error("Error reading preset " + error.toString(), error.stack)
-        .finally -> done()
-      , ->
-        deferred.resolve(fileContents)
-      )
-    .catch (error) ->
-      logger.error(error, error.stack)
-    deferred.promise
-
   # read available presets from zookeeper
   #
   # @return [Promise] resolves with list of presets
@@ -63,7 +33,6 @@ class ModulePreset
 
   # determines diff between return of presetSourceMethod and active zookeeper storage and
   # creates an array of presets, that have not yet been added to zookeeper
-  #
   # @param [Function] presetSourceMethod has to be a promise returning method, that resolves with an array
   # of the presets [{appId: dev-app-1, options:{}...},{appId: dev-app-2 ...]
   #
@@ -75,21 +44,25 @@ class ModulePreset
     presetSourceMethod()
     .then (sourceContents) ->
       pushToZookeeper = []
-      async.eachLimit(sourceContents,1, (sc, done) ->
-        appid = sc.appId.split("/")[1]
-        ZookeeperHandler.exists("modules/#{sc.moduleName}/presets/#{appid}")
-        .then (exists) ->
-          if allwaysUpdate and exists
-            logger.info "Update preset \"#{sc.appId}\" for module \"#{sc.moduleName}\""
-            sc.lastEdit = new Date().getTime()
-            pushToZookeeper.push(sc)
-          else unless exists
-            logger.info "Create preset \"#{sc.appId}\"for module \"#{sc.moduleName}\""
-            sc.lastEdit = false
-            pushToZookeeper.push(sc)
-        .catch (error) ->
-          logger.error("Create error \"" + error + "\"", error.stack)
-        .finally ->
+      async.eachLimit(sourceContents, 1, (sc, done) ->
+        if sc.appId?
+          appid = sc.appId.split("/")[1]
+          ZookeeperHandler.exists("modules/#{sc.moduleName}/presets/#{appid}")
+          .then (exists) ->
+            if allwaysUpdate and exists
+              logger.info "Update preset \"#{sc.appId}\" for module \"#{sc.moduleName}\""
+              sc.lastEdit = new Date().getTime()
+              pushToZookeeper.push(sc)
+            else unless exists
+              logger.info "Create preset \"#{sc.appId}\" for module \"#{sc.moduleName}\""
+              sc.lastEdit = false
+              pushToZookeeper.push(sc)
+          .catch (error) ->
+            logger.error("Create error \"" + error + "\"", error.stack)
+          .finally ->
+            done()
+        else
+          logger.error("Error appId is invalid set", "")
           done()
       , ->
         deferred.resolve(pushToZookeeper)
@@ -109,42 +82,53 @@ class ModulePreset
   @sync: (presetSourceMethod, allwaysUpdate = false) ->
     deferred = Q.defer()
     errors = []
+    messages = []
     count = 0
     @diff(presetSourceMethod, allwaysUpdate)
     .then (result) =>
       async.each(result,
         (preset, done) =>
-          clearedAppName = preset.appId.split("/")[1]
-          Module = require("./Module")
-          unless Module.isEnabled(preset.moduleName)
-            message = "Module \"#{preset.moduleName}\" is not enabled - skipping preset for app \"#{clearedAppName}\""
-            logger.error(message, "")
-            errors.push(message)
-            done()
-          else
-            count++
-            if preset.lastEdit
-              @editPreset(preset, clearedAppName, done)
+          if preset.appId?
+            clearedAppName = preset.appId.split("/")[1]
+            Module = require("./Module")
+            unless Module.isEnabled(preset.moduleName)
+              logger.error("Module not found: #{preset.moduleName}", "")
+              errors.push("Module not found: #{preset.moduleName}")
+              done()
             else
-              @createPreset(preset, clearedAppName, done)
-        , ->
+              count++
+              if preset.lastEdit
+                @editPreset(preset, clearedAppName, messages, done)
+              else
+                @createPreset(preset, clearedAppName, messages, done)
+          else
+            logger.error("Error appId is invalid set", "")
+            done()
+      , ->
+        if errors.count or count is 0
           deferred.resolve(
-            count: count
-            errors: errors
+            status: "error"
+            message: errors.join(", ")
+          )
+        else
+          deferred.resolve(
+            status: "ok"
+            message: messages.join("")
           )
       )
     .catch (error) ->
       logger.error(error, error.stack)
     deferred.promise
 
-  @editPreset: (preset, clearedAppName, done) ->
+  @editPreset: (preset, clearedAppName, messages, done) ->
     deferred = Q.defer()
     logger.info("Forward preset update \"#{clearedAppName}\" for module \"#{preset.moduleName}\" to
                 ZookeeperHandler")
-    ZookeeperHandler.setData("modules/#{preset.moduleName}/presets/#{clearedAppName}",preset)
+    ZookeeperHandler.setData("modules/#{preset.moduleName}/presets/#{clearedAppName}", preset)
     .then ->
       Module.modules[preset.moduleName].updatePresetsOnGui()
       Module.modules[preset.moduleName].editPreset(preset)
+      messages.push "AppId for module #{preset.moduleName} modified: /#{clearedAppName}"
       deferred.resolve()
     .catch (error) ->
       logger.error(error, error.stack)
@@ -153,7 +137,7 @@ class ModulePreset
       done()
     deferred.promise
 
-  @createPreset: (preset, clearedAppName, done) ->
+  @createPreset: (preset, clearedAppName, messages, done) ->
     deferred = Q.defer()
     logger.info("Forward preset creation \"#{clearedAppName}\" for module \"#{preset.moduleName}\" to
                 ZookeeperHandler")
@@ -162,6 +146,7 @@ class ModulePreset
     .then ->
       Module.modules[preset.moduleName].updatePresetsOnGui()
       Module.modules[preset.moduleName].createPreset(preset)
+      messages.push "AppId for module #{preset.moduleName} created: /#{clearedAppName}"
       deferred.resolve()
     .catch (error) ->
       deferred.reject(error)
@@ -178,23 +163,37 @@ class ModulePreset
   @remove: (presets) ->
     deferred = Q.defer()
     errors = []
+    messages = []
     async.each(presets,
       (preset, done) ->
-        clearedAppName = preset.appId.split("/")[1]
-        logger.warn("Remove preset \"" + preset.appId + "\" for module \"#{preset.moduleName}\"")
-        ZookeeperHandler.remove("modules/#{preset.moduleName}/presets/#{clearedAppName}")
-        .then ->
-          Module.modules[preset.moduleName].deletePreset(preset)
-        .catch (error) ->
-          message = "Error removing preset #{preset.appId} for  \"" + preset.moduleName + "\" \"" + error.toString() +  "\": not found"
-          errors.push message
-          logger.error(message, "")
-        .finally ->
+        if preset.appId?
+          clearedAppName = preset.appId.split("/")[1]
+          logger.info("Remove preset \"" + preset.appId + "\" for module \"#{preset.moduleName}\"")
+          ZookeeperHandler.remove("modules/#{preset.moduleName}/presets/#{clearedAppName}")
+          .then ->
+            Module.modules[preset.moduleName].deletePreset(preset)
+            Module.modules[preset.moduleName].updatePresetsOnGui()
+            messages.push "AppId for module #{preset.moduleName} deleted: /#{clearedAppName}"
+          .catch (error) ->
+            message = "Error removing preset #{preset.appId} for  \"" + preset.moduleName + "\" \"" + error.toString() +  "\": not found"
+            errors.push message
+            logger.error(message, "")
+          .finally ->
+            done()
+        else
+          logger.error("Error appId is invalid set", "")
           done()
       , ->
-        deferred.resolve(
-          errors: errors
-        )
+        if errors.length
+          deferred.resolve(
+            status: "error"
+            message: errors.join(", ")
+          )
+        else
+          deferred.resolve(
+            status: "ok"
+            message: messages.join(", ")
+          )
     )
     deferred.promise
 
@@ -259,7 +258,7 @@ class ModulePreset
   @getAllOfModule: (moduleName) ->
     deferred = Q.defer()
     unless Module.modules[moduleName]?
-      deferred.reject("module #{moduleName} not active or unknown")
+      deferred.reject("Module not found: \"#{moduleName}\"")
     else
       deferred.resolve(Module.modules[moduleName].presets)
     deferred.promise
